Update getFavorites for using search parameter 'ids' of the punkApi
